# Rewrite Uri.CheckForUnicode to improve throughput and reduce allocations